Move email function in new module
