+++ src/gtimelog/__init__.py
@@ -1,3 +1,8 @@
 # The gtimelog package.
 
+import sys
+
+
 __version__ = '0.12.0.dev0'
+
+DEBUG = '--debug' in sys.argv
